+++ src/pages/Schedule/Week/TeamsOnBye/TeamsOnBye.styled.jsx
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const TeamsOnByeContainer = styled.div`
+  text-align: left;
+  padding: 15px 0 0 0;
+  margin: auto 10%;
+  font-size: 12px;
+  color: #4c8bcf;
+  @media (min-width: 901px) {
+    padding-left: 0;
+    text-align: center;
+  }
+`;
