+++ src/pages/Schedule/Week/WeekSelected.jsx
@@ -1,44 +1,9 @@
 import { ApiCalls } from "../../../utils/apiCalls";
-import Game from "../../../components/Game/Game";
-import styled from "styled-components";
-
-import { CircularProgress } from "@mui/material";
 import { groupByDayOfTheWeek } from "../../../utils/cleanUpDataHelpers";
-
-const S = {};
-S.TeamsOnByeContainer = styled.div`
-  text-align: left;
-  padding: 15px 0 0 0;
-  margin: auto 10%;
-  font-size: 12px;
-  color: #4c8bcf;
-  @media (min-width: 901px) {
-    padding-left: 0;
-    text-align: center;
-  }
-`;
-S.AllGamesContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  flex-wrap: wrap;
-  justify-content: center;
-  margin-top: 20px;
-`;
-S.DayHeading = styled.h4`
-  padding-left: 9vw;
-  margin: 25px 0 15px 0;
-  color: #013369;
-  @media (min-width: 901px) {
-    padding-left: 0;
-    text-align: center;
-  }
-`;
-S.GamesByDayGroup = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`;
-
+import TeamsOnBye from "./TeamsOnBye/TeamsOnBye";
+import GameCard from "./GameCard/GameCard";
+import { CircularProgress } from "@mui/material";
+import * as S from "./WeekSelected.styled";
 
 export default function WeekSelected({ weekNo }) {
 
@@ -55,22 +20,8 @@ export default function WeekSelected({ weekNo }) {
 
   return (
     <div style={{ minHeight: "90vh" }}>
-      <S.TeamsOnByeContainer>
-        {!teamsOnBye.isLoading && teamsOnBye.data.length !== 0 ? (
-          teamsOnBye.isLoading ? (
-            <CircularProgress />
-          ) : (
-            <div>
-              <span>TEAMS ON BYE (WEEK {weekNo}): </span>
-              {teamsOnBye.data.map((team, i, arr) => {
-                return (
-                  <span key={`team-on-bye-${team.id}`} >{team.displayName.toUpperCase()}{i !== arr.length - 1 ? ', ' : ''} </span>
-                )
-              })}
-            </div>
-          )
-        ) : (weekNo > 18 ? "" : <em>NO TEAMS ON BYE FOR SELECTED WEEK</em>)}
-      </S.TeamsOnByeContainer>
+
+      <TeamsOnBye weekNo={weekNo} teamsOnBye={teamsOnBye} />
 
       {
         allGamesData.isLoading ? (
@@ -88,11 +39,10 @@ export default function WeekSelected({ weekNo }) {
                   <S.GamesByDayGroup>
                     {groupGamesByDayOfTheWeek[key].map(game => {
                       return (
-                        <Game
+                        <GameCard
                           event={game}
                           key={`${game.id}-${Math.random().toString()}`}
                           weekNo={weekNo}
-                          type="week-list-game"
                         />
                       )
                     })}
