+++ src/pages/Team/TeamTabs/TabsContent/GameCardForTeam.styled.jsx
